src/python/requirements.txt
kafka-python==2.0.2
chromadb==0.5.3
sentence-transformers==2.7.0
numpy==1.26.4
